feat: Refactor Go packages to ease collaboration
