+++ src/autodetect/main.go
