+++ src/instrumentation/podmutator.go
@@ -25,8 +25,8 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/newrelic/k8s-agents-operator/api/v1alpha1"
-	"github.com/newrelic/k8s-agents-operator/internal/webhookhandler"
+	"github.com/newrelic/k8s-agents-operator/src/api/v1alpha1"
+	"github.com/newrelic/k8s-agents-operator/src/internal/webhookhandler"
 )
 
 var (
