+++ src/instrumentation/sdk.go
@@ -36,8 +36,8 @@ import (
 	"k8s.io/client-go/util/retry"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/newrelic/k8s-agents-operator/api/v1alpha1"
-	"github.com/newrelic/k8s-agents-operator/pkg/constants"
+	"github.com/newrelic/k8s-agents-operator/src/api/v1alpha1"
+	"github.com/newrelic/k8s-agents-operator/src/constants"
 )
 
 const (
