+++ src/internal/config/main_test.go
@@ -24,8 +24,8 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/newrelic/k8s-agents-operator/internal/config"
-	"github.com/newrelic/k8s-agents-operator/pkg/autodetect"
+	"github.com/newrelic/k8s-agents-operator/src/autodetect"
+	"github.com/newrelic/k8s-agents-operator/src/internal/config"
 )
 
 func TestNewConfig(t *testing.T) {
