+++ src/internal/config/options.go
@@ -23,8 +23,8 @@ import (
 
 	"github.com/go-logr/logr"
 
-	"github.com/newrelic/k8s-agents-operator/internal/version"
-	"github.com/newrelic/k8s-agents-operator/pkg/autodetect"
+	"github.com/newrelic/k8s-agents-operator/src/autodetect"
+	"github.com/newrelic/k8s-agents-operator/src/internal/version"
 )
 
 // Option represents one specific configuration option.
