+++ src/internal/version/main.go
