+++ src/internal/version/main_test.go
